requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ML / NLP
torch>=2.0.0
transformers>=4.40.0
tokenizers>=0.19.0

# NLP Preprocessing
spacy>=3.7.0
nltk>=3.8.0

# Document Parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# Data
pydantic>=2.0.0

# CORS
starlette>=0.38.0
